+++ lib/roo_on_rails/sidekiq/settings.rb
@@ -3,7 +3,6 @@ require_relative './queue_latency'
 module RooOnRails
   module Sidekiq
     class Settings
-
       DEFAULT_QUEUE_LATENCY_VALUES = {
         'monitoring' => 10.seconds.to_i,
         'realtime' => 10.seconds.to_i,
@@ -15,12 +14,39 @@ module RooOnRails
         'default' => 1.day.to_i
       }.freeze
 
-      def self.queues
-        @queues ||= QueueLatency.permitted_latency_values.sort_by(&:last).map(&:first).freeze
-      end
+      class << self
+        def queues
+          @queues ||= permitted_latency_values.sort_by(&:last).map(&:first).freeze
+        end
 
-      def self.concurrency
-        ENV.fetch('SIDEKIQ_THREADS', 25)
+        def concurrency
+          ENV.fetch('SIDEKIQ_THREADS', 25)
+        end
+
+        def permitted_latency_values
+          @permitted_latency_values ||= ENV.key?('SIDEKIQ_QUEUES') ? env_queue_latency_values.freeze : DEFAULT_QUEUE_LATENCY_VALUES
+        end
+
+        private
+
+        def env_queue_latency_values
+          {}.tap do |hash|
+            ENV['SIDEKIQ_QUEUES'].split(',').each do |entry|
+              queue_entry = entry.strip
+              if DEFAULT_QUEUE_LATENCY_VALUES.key?(queue_entry)
+                queue_name = queue_entry
+                hash[queue_name] = DEFAULT_QUEUE_LATENCY_VALUES[queue_entry]
+              elsif queue_entry.match(/\Awithin\d+.+\z/)
+                _, number, unit = queue_entry.partition(/\d+/)
+                hash[queue_entry] = number.to_i.public_send(unit.strip).to_i
+              elsif queue_entry.include?(':')
+                queue_name, latency_info = queue_entry.split(':')
+                _, number, unit = latency_info.partition(/\d+/)
+                hash[queue_name] = number.to_i.public_send(unit.strip).to_i
+              end
+            end
+          end
+        end
       end
     end
   end
